Widen token IDs to 116 bits with u64 species and encoded tier/type
First step toward permissionless community species registration
(docs/community-beasts-design.md): the token ID format must support
unbounded species IDs and carry all static beast data before the
registry can exist.

- PackableBeast.id widens to u64; tier (3 bits) and type (3 bits) are
  now encoded in the token ID, resolved by the contract at mint time
  and never caller-supplied, making get_beast_power a pure function of
  the token ID
- decode_token_id validates format-level invariants: non-zero id,
  tier 1-5, type <= 2, affix ranges, and both-or-neither zero affixes;
  legacy 53-bit token IDs are rejected
- Non-genesis mints now require prefix >= 1 and suffix >= 1, reserving
  the (id, 0, 0) slot for each species' Genesis Beast and capping every
  species at exactly 1,243 tokens; genesis status derives from the ID
- beast_id widens to u64 across interfaces, ranking maps, minting
  coordinator, and metadata; beast hashes are unchanged for species
  1-75 (poseidon over felts), preserving Death Mountain history
- Attach previously-orphaned mint_tests.cairo and tests.cairo as test
  modules, fix their stale OZ imports and base64 token_uri assertions

+++ src/interfaces.cairo
@@ -14,7 +14,7 @@ pub trait IBeasts<TContractState> {
     fn mint(
         ref self: TContractState,
         to: ContractAddress,
-        beast_id: u8,
+        beast_id: u64,
         prefix: u8,
         suffix: u8,
         level: u16,
@@ -24,20 +24,20 @@ pub trait IBeasts<TContractState> {
     ) -> (u256, u16, bool);
 
     // Metadata functions
-    fn refresh_metadata(ref self: TContractState, beast_id: u8);
+    fn refresh_metadata(ref self: TContractState, beast_id: u64);
     fn refresh_dungeon_stats(ref self: TContractState, token_id: u256);
 
     // Beast queries
     fn get_beast(self: @TContractState, token_id: u256) -> PackableBeast;
-    fn is_minted(self: @TContractState, beast_id: u8, prefix: u8, suffix: u8) -> bool;
+    fn is_minted(self: @TContractState, beast_id: u64, prefix: u8, suffix: u8) -> bool;
     fn total_supply(self: @TContractState) -> u256;
 
     // Beast ranking queries
     fn get_beast_rank(self: @TContractState, token_id: u256) -> u16;
-    fn get_token_id_at_rank(self: @TContractState, beast_id: u8, rank: u16) -> u256;
-    fn get_species_count(self: @TContractState, beast_id: u8) -> u16;
+    fn get_token_id_at_rank(self: @TContractState, beast_id: u64, rank: u16) -> u256;
+    fn get_species_count(self: @TContractState, beast_id: u64) -> u16;
     fn get_kill_count(self: @TContractState, token_id: u256) -> u64;
-    fn get_beast_metadata_bookmark(self: @TContractState, beast_id: u8) -> u16;
+    fn get_beast_metadata_bookmark(self: @TContractState, beast_id: u64) -> u16;
     fn get_last_manual_metadata_refresh(self: @TContractState, token_id: u256) -> u64;
     fn get_adventurer_killed(self: @TContractState, token_id: u256, index: u64) -> u64;
     fn get_last_killed_timestamp(self: @TContractState, token_id: u256) -> u64;
